storage: move persist workers into their own module
